Get rtcp-fb tests working again.
The problem here was that, prior to our rewrite, sipcc never tried to
use more than one codec on any given m-line, which hid an rtcp-fb
handling bug in the webrtc.org code. This code configured rtcp-fb on a
per-conduit basis, while rtcp-fb can be specified on a per-codec basis
in the SDP. We may need to disable the multi-recv-codec case, since
we've never actually exercised those codepaths before.